# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Callcenter",
        "version": "1.0"
    },
    "paths": {
...
        "/api/pais": {
            "get": {
                "tags": [
                    "Pais"
                ],
                "summary": "Lista todos los Paises",
                "description": "Retorna todos los paises",
                "operationId": "getPaisList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Pais": {
                                            "description": "Listado de países",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "IDPais": {
                                                        "description": "ID de País",
                                                        "type": "integer"
                                                    },
                                                    "Pais": {
                                                        "description": "Nombre de País",
                                                        "type": "string"
                                                    },
                                                    "BaseFactura": {
                                                        "description": "Monto para consumidor final",
                                                        "type": "integer"
                                                    },
                                                    "Cadenas": {
                                                        "description": "Cadenas pertenecientes al País",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "IDCadena": {
                                                                    "description": "ID de la Cadena",
                                                                    "type": "integer"
                                                                },
                                                                "Cadena": {
                                                                    "description": "Nombre de la Cadena",
                                                                    "type": "string"
                                                                },
                                                                "ServicioDomicilio": {
                                                                    "description": "Valor Servicio a Domicilio",
                                                                    "type": "integer"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "Impuestos": {
                                                        "description": "Impuestos aplicados al País",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "IDImpuesto": {
                                                                    "description": "ID del Impuesto",
                                                                    "type": "string"
                                                                },
                                                                "Descripcion": {
                                                                    "description": "Descripción del Impuesto",
                                                                    "type": "string"
                                                                },
                                                                "Porcentaje": {
                                                                    "description": "Porcentaje del Impuesto",
                                                                    "type": "integer"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "Ciudades": {
                                                        "description": "Ciudades del País",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "IDCiudad": {
                                                                    "description": "ID de la Ciudad",
                                                                    "type": "integer"
                                                                },
                                                                "Ciudad": {
                                                                    "description": "Nombre de la Ciudad",
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/restaurante/IDRestaurante/{IDRestaurante}": {
            "get": {
                "tags": [
                    "Restaurante"
                ],
                "summary": "Lista restaurantes por IDRestaurante",
                "description": "Retorna restaurante por IDRestaurante",
                "operationId": "getRestauranteId",
                "parameters": [
                    {
                        "name": "IDRestaurante",
                        "in": "path",
                        "description": "Puede buscar por uno ó varios Ids separados por coma /4 ó /4,20",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/restaurante/IDCadena/{IDCadena}": {
            "get": {
                "tags": [
                    "Restaurante"
                ],
                "summary": "Lista restaurantes por IDCadena",
                "description": "Retorna restaurante por IDCadena",
                "operationId": "getRestaurantePorIDCadena",
                "parameters": [
                    {
                        "name": "IDCadena",
                        "in": "path",
                        "description": "Puede buscar por uno ó varios Ids separados por coma /10 ó /12,10",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}